Rename routes to make it more RESTful
