+++ server/routes/label_route.js
@@ -1,22 +1,15 @@
 const router = require("express").Router();
-const { upload, uploadS3, authentication } = require("../../util/util");
-
-// const oriUpload = upload.single("OriImage");
-const oriUploadS3 = uploadS3.single("OriImage");
+const { authentication } = require("../../util/util");
 
 const {
-  saveOriImage,
   saveCoordinates,
-  loadLabels
+  getLabels
 } = require("../controllers/label_controller");
-
-router.route("/label/ori-image")
-  .post(authentication(), oriUploadS3, saveOriImage);
 
 router.route("/label/coordinates")
   .post(authentication(), saveCoordinates);
 
-router.route("/label/load-coordinates")
-  .get(loadLabels);
+router.route("/label/coordinates")
+  .get(getLabels);
 
 module.exports = router;
